Fixed fortran 90 regression test problems to properly handle int vs long int arguments between ARKode solver and legacy fortran 77 code; results in a slight performance improvement for a couple of tests, but should really help in ensuring results across different platforms and compilers (so long as they're all 64-bit).

+++ regression/arkode/ark_bruss.f90
@@ -54,8 +54,8 @@ program driver
   !    rpar(1) -> "a" parameter
   !    rpar(2) -> "b" parameter 
   !    rpar(3) -> "ep" parameter
-  integer :: ipar
-  real*8  :: rpar(3)
+  integer*8 :: ipar
+  real*8    :: rpar(3)
 
   ! solver parameters
   logical :: denseout
@@ -245,11 +245,11 @@ subroutine farkifun(t, y, ydot, ipar, rpar, ier)
   implicit none
 
   ! Arguments
-  real*8,  intent(in)  :: t, rpar(3)
-  integer, intent(in)  :: ipar(1)
-  integer, intent(out) :: ier
-  real*8,  intent(in)  :: y(3)
-  real*8,  intent(out) :: ydot(3)
+  real*8,  intent(in)   :: t, rpar(3)
+  integer*8, intent(in) :: ipar(1)
+  integer, intent(out)  :: ier
+  real*8,  intent(in)   :: y(3)
+  real*8,  intent(out)  :: ydot(3)
 
   ! temporary variables
   real*8 :: u, v, w, a, b, ep
@@ -287,11 +287,11 @@ subroutine farkefun(t, y, ydot, ipar, rpar, ier)
   implicit none
 
   ! Arguments
-  real*8,  intent(in)  :: t, rpar(3)
-  integer, intent(in)  :: ipar(1)
-  integer, intent(out) :: ier
-  real*8,  intent(in)  :: y(3)
-  real*8,  intent(out) :: ydot(3)
+  real*8,  intent(in)   :: t, rpar(3)
+  integer*8, intent(in) :: ipar(1)
+  integer, intent(out)  :: ier
+  real*8,  intent(in)   :: y(3)
+  real*8,  intent(out)  :: ydot(3)
 
   ! temporary variables
   real*8 :: u, v, w, a, b, ep
@@ -329,11 +329,11 @@ subroutine farkdjac(neq,t,y,fy,DJac,h,ipar,rpar,wk1,wk2,wk3,ier)
   implicit none
 
   ! Arguments
-  real*8,  intent(in)  :: t, h, rpar(3)
-  integer, intent(in)  :: neq, ipar(1)
-  integer, intent(out) :: ier
+  real*8,  intent(in)   :: t, h, rpar(3)
+  integer*8, intent(in) :: neq, ipar(1)
+  integer, intent(out)  :: ier
   real*8,  intent(in), dimension(neq) :: y, fy, wk1, wk2, wk3
-  real*8,  intent(out) :: DJac(neq,neq)
+  real*8,  intent(out)  :: DJac(neq,neq)
 
   ! temporary variables
   real*8 :: u, v, w, a, b, ep
